+++ app/src/main/java/com/olderwold/reddit/MainActivity.kt
@@ -6,9 +6,9 @@ import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
-import com.olderwold.reddit.feed.WebPage
 import com.olderwold.reddit.ui.element.RedditHotList
 import com.olderwold.reddit.ui.theme.YetanotherredditTheme
+import com.olderwold.reddit.web.WebPage
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
